+++ packages/components-js/tests/e2e/specs/examples.e2e.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
-import * as path from 'path';
 import * as fs from 'fs';
+import * as path from 'path';
 import {
   getConsoleErrorMessages,
   getConsoleErrorsAmount,
@@ -35,7 +35,14 @@ for (const url of exampleUrls) {
   test(`if example at ${url} works without error or warning`, async ({ page, browserName }) => {
     test.skip(browserName !== 'chromium');
     // Skip AG Grid pages since they will show licensing errors
-    test.skip(['ag-grid-example', 'ag-grid-example-storefront'].includes(url));
+    test.skip(
+      [
+        'ag-grid-example',
+        'ag-grid-example-storefront',
+        'ag-grid-example-compact',
+        'ag-grid-example-compact-storefront',
+      ].includes(url)
+    );
     await goto(page, url);
 
     if (getConsoleErrorsAmount() !== 0) {